REFACTOR: orders API implementation into ordersController
